feat: Implement Add, View, and Edit Workout functionality with media and URL support
- Added AddWorkout component for creating and editing workouts, including media capture and URL links.
- Introduced ViewWorkout component to display workout details, including sets, media, and external links.
- Enhanced WorkoutsPageContent to handle different modes (add, edit, view) and manage workout state.
- Implemented search and category filtering for workouts.
- Added toast notifications for user feedback on actions like saving and deleting workouts.
