+++ package.json
@@ -2,6 +2,7 @@
   "name": "gym-manager-app",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
