+++ public/fallback-RElg5vYlmoF8df6nDlylp.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";self.fallback=async e=>{switch(e.destination){case"document":return caches.match("/offline",{ignoreSearch:!0});case"image":return caches.match("/icons/icon-192x192.png",{ignoreSearch:!0});default:return Response.error()}}})();
