+++ public/sw.js
@@ -1,0 +1,1 @@
+if(!self.define){let e,s={};const n=(n,a)=>(n=new URL(n+".js",a).href,s[n]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()}).then(()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e}));self.define=(a,i)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const l=e=>n(e,c),r={module:{uri:c},exports:t,require:l};s[c]=Promise.all(a.map(e=>r[e]||l(e))).then(e=>(i(...e),t))}}define(["./workbox-b9a5f6f7"],function(e){"use strict";importScripts("fallback-RElg5vYlmoF8df6nDlylp.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"3f60811261f141174f1308345e1fe2cb"},{url:"/_next/static/RElg5vYlmoF8df6nDlylp/_buildManifest.js",revision:"013bfd9c6837d26ef190778df4b9a5ca"},{url:"/_next/static/RElg5vYlmoF8df6nDlylp/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/185-2b133febe16d029f.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/290.112ae66ce474890f.js",revision:"112ae66ce474890f"},{url:"/_next/static/chunks/355.e5d8bea266c62028.js",revision:"e5d8bea266c62028"},{url:"/_next/static/chunks/37.5ca6b45cc23a3637.js",revision:"5ca6b45cc23a3637"},{url:"/_next/static/chunks/412-022426678b0e7aa5.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/4bd1b696-26d13eb3515be7a9.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/526-e9d31b70faeb24b5.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/597-d6c9a1c1e8dd4cd7.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/662-ad61b436ef456e29.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/684-79069bace43cbc62.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/736.e5d8bea266c62028.js",revision:"e5d8bea266c62028"},{url:"/_next/static/chunks/770-187fa55991ed79df.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/82-5fb630b16900226e.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/822-6b1ced3d1c36cf45.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/970-0cc47cbdde84a96c.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/_not-found/page-fc4faa4fb996c7f5.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/api/url-metadata/route-3bdd852d9cc195bb.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/error-70d094b7d0a6dda8.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/layout-2fdbefe05edb0d42.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/offline/page-4ac6c70b62f7045d.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/page-167b299193fbde68.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/robots.txt/route-f3ff6372000fa082.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/sitemap.xml/route-53412e0f59e4f95f.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/workouts/layout-ba42783af29b6ed5.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/app/workouts/page-269afda4ac6e49ea.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/framework-75d8b6119e7601c4.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/main-app-52e271e6034b02b7.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/main-bf1a474e35827158.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/pages/_app-8e94039938385921.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/pages/_error-7b2d139042a6a5ab.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-5c49fc33a4596a38.js",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/_next/static/css/6bdb8f8f3cbeb297.css",revision:"6bdb8f8f3cbeb297"},{url:"/_next/static/css/78034be178d05145.css",revision:"78034be178d05145"},{url:"/_next/static/css/df630674007affb9.css",revision:"df630674007affb9"},{url:"/_next/static/media/046b90749014f852-s.woff2",revision:"19bf2a23f7f672153135a9d1918f6f9a"},{url:"/_next/static/media/19cfc7226ec3afaa-s.woff2",revision:"9dda5cfc9a46f256d0e131bb535e46f8"},{url:"/_next/static/media/21350d82a1f187e9-s.woff2",revision:"4e2553027f1d60eff32898367dd4d541"},{url:"/_next/static/media/67110d8fe39c5e84-s.woff2",revision:"91c073ec3046c2fc252900a89b6fc5d0"},{url:"/_next/static/media/6aacc40b7795b725-s.woff2",revision:"48e07fe2ca9c3bc32d09affb2ace8844"},{url:"/_next/static/media/848b99572ad207f3-s.woff2",revision:"31904e07bc2fac21149accd8a82eb1b1"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/999e639cd9d85971-s.woff2",revision:"59533f46ae2b6e4fed5c133c03ea0608"},{url:"/_next/static/media/ba9851c3c22cd980-s.woff2",revision:"9e494903d6b0ffec1a1e14d34427d44d"},{url:"/_next/static/media/c5fe6dc8356a8c31-s.woff2",revision:"027a89e9ab733a145db70f09b8a18b42"},{url:"/_next/static/media/c97d4358b5ad6f1f-s.p.woff2",revision:"748da8fce84b0b6ee83bacd60aed2979"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/_next/static/media/e6b5cfd5a74e1cae-s.woff2",revision:"8358e3d9b140dd03a59878681e98a5e4"},{url:"/android-chrome-192x192.png",revision:"9c18adbf7c657ee767efaf4c079030c3"},{url:"/android-chrome-512x512.png",revision:"1c160e01286a67ba08bbae9cbb78f8bc"},{url:"/apple-touch-icon.png",revision:"d6f1322f54f1e729d3561cf899af2919"},{url:"/favicon-16x16.png",revision:"ffee7f415d1e6fd01666b44c53fb411a"},{url:"/favicon-32x32.png",revision:"8f5899d20331b7773693e452c60afa64"},{url:"/favicon.ico",revision:"2bdc42994e2cf08bb4cc95815b93e555"},{url:"/icons/add-shortcut.png",revision:"cf489d6e11865da09690420cb983ba9f"},{url:"/icons/icon-128x128.png",revision:"30688ac09f1a589eac29295142518aaa"},{url:"/icons/icon-144x144.png",revision:"30688ac09f1a589eac29295142518aaa"},{url:"/icons/icon-152x152.png",revision:"30688ac09f1a589eac29295142518aaa"},{url:"/icons/icon-192x192.png",revision:"30688ac09f1a589eac29295142518aaa"},{url:"/icons/icon-324x324.png",revision:"6d97dd0e1c6f7e27418166fb8a016dbf"},{url:"/icons/icon-384x384.png",revision:"30688ac09f1a589eac29295142518aaa"},{url:"/icons/icon-512x512.png",revision:"cf489d6e11865da09690420cb983ba9f"},{url:"/icons/icon-72x72.png",revision:"7834a1b7a5394b58a2c240effdc9ad6e"},{url:"/icons/icon-86x86.png",revision:"e0ae51d67ec49f68323dfd390006b9a1"},{url:"/icons/icon-96x96.png",revision:"bdb046298089430a0dff1be2fb9a73f3"},{url:"/icons/icon-base.svg",revision:"9c0491dc576d1c16208c132b66fee698"},{url:"/icons/screenshot-1.png",revision:"f91f0757a45d0d85111badac3356a565"},{url:"/icons/screenshot-2.png",revision:"f91f0757a45d0d85111badac3356a565"},{url:"/icons/workouts-shortcut.png",revision:"cf489d6e11865da09690420cb983ba9f"},{url:"/manifest.json",revision:"17fdf20513f83c1f2b05ba9ce861da9c"},{url:"/offline",revision:"RElg5vYlmoF8df6nDlylp"},{url:"/robots.txt",revision:"c705e80aec228cdbe7a3b7fd28e1ef08"},{url:"/site.webmanifest",revision:"053100cb84a50d2ae7f5492f7dd7f25e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:a})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[new e.ExpirationPlugin({maxEntries:1,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js|css|woff|woff2|ttf|eot)$/i,new e.CacheFirst({cacheName:"static-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:png|jpg|jpeg|svg|gif|webp|ico)$/i,new e.CacheFirst({cacheName:"images",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https?:\/\/[^\/]+\/(?!api\/).*/,new e.NetworkFirst({cacheName:"pages-cache",networkTimeoutSeconds:3,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),new e.CacheableResponsePlugin({statuses:[0,200]}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.googleapis\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.gstatic\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-static",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")});
