+++ src/app/robots.ts
@@ -1,7 +1,7 @@
 import { MetadataRoute } from 'next'
 
 export default function robots(): MetadataRoute.Robots {
-  const baseUrl = process.env.NEXT_PUBLIC_BASE_URL || 'https://localhost:3000'
+  const baseUrl = 'https://workout-manager-app.ahemdsoran.dev'
 
   return {
     rules: {
