+++ src/app/sitemap.ts
@@ -19,7 +19,7 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: 0.9,
     },
     {
-      url: `${baseUrl}/workouts/add`,
+      url: `${baseUrl}/workouts?action=add`,
       lastModified: new Date(),
       changeFrequency: 'weekly',
       priority: 0.8,
