+++ src/components/navbar.tsx
@@ -34,7 +34,7 @@ export default function Navbar() {
                 href={item.href}>
                 {item.href === '/' && <IconHome size={18} />}
                 {item.href === '/workouts' && <IconList size={18} />}
-                {item.href === '/workouts/add' && <IconPlus size={18} />}
+                {item.href === '/workouts?action=add' && <IconPlus size={18} />}
                 {item.label}
               </Link>
             </NavbarItem>
@@ -66,7 +66,7 @@ export default function Navbar() {
                 size='lg'>
                 {item.href === '/' && <IconHome size={20} />}
                 {item.href === '/workouts' && <IconList size={20} />}
-                {item.href === '/workouts/add' && <IconPlus size={20} />}
+                {item.href === '/workouts?action=add' && <IconPlus size={20} />}
                 {item.label}
               </Link>
             </NavbarMenuItem>
