+++ tsconfig.json
@@ -14,6 +14,8 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
+    "noUnusedLocals": true,
+    "noImplicitAny": true,
     "plugins": [
       {
         "name": "next"
